+++ NEWS.md
@@ -1,3 +1,11 @@
+# gips 1.2.1.9000
+
+### Bugfix:
+
+- `logLik.gips()` will return an object of class `logLik`;
+- Better Vignettes titles.
+
+
 # gips 1.2.1
 
 ### Bugfix: